feat: refactored infinite scroll to custom hook
